Moved brand ramp into styles
