+++ samples/apps/copilot-chat-app/webapp/src/index.tsx
@@ -12,7 +12,7 @@ import { store } from './redux/app/store';
 
 import React from 'react';
 import { getMissingEnvVariables } from './checkEnv';
-import { BrandVariants, createLightTheme } from '@fluentui/react-components';
+import { semanticKernelLightTheme } from './styles';
 
 
 if (!localStorage.getItem('debug')) {
@@ -21,26 +21,6 @@ if (!localStorage.getItem('debug')) {
 
 let container: HTMLElement | null = null;
 
-const semanticKernelBrandRamp: BrandVariants = {
-    10: "#060103",
-    20: "#261018",
-    30: "#431426",
-    40: "#591732",
-    50: "#701A3E",
-    60: "#861F4B",
-    70: "#982C57",
-    80: "#A53E63",
-    90: "#B15070",
-    100: "#BC627E",
-    110: "#C6748B",
-    120: "#CF869A",
-    130: "#D898A8",
-    140: "#E0AAB7",
-    150: "#E8BCC6",
-    160: "#EFCFD6"
-};
-
-export const semanticKernelLightTheme = createLightTheme(semanticKernelBrandRamp);
 
 document.addEventListener('DOMContentLoaded', async () => {
     if (!container) {
